requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1.1
django-cors-headers==3.5.0
django-rest-framework==0.1.0
djangorestframework==3.11.1
gunicorn==20.0.4
idna==2.10
numpy==1.19.1
pandas==1.1.1
postgres==3.0.0
prompt-toolkit==3.0.7
psycopg2-binary==2.8.6
psycopg2-pool==1.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.19
sqlparse==0.3.1
urllib3==1.25.10
wcwidth==0.2.5
whitenoise==5.2.0
wit==6.0.0
